requirements.txt
opencv-python-headless>=4.8
Pillow>=10.0
numpy>=1.24
pandas>=2.0
scikit-learn>=1.3
xgboost>=2.0
shap>=0.44
imagehash>=4.3
torch>=2.0
transformers>=4.35
accelerate>=0.24
mediapipe>=0.10.18
streamlit>=1.28
plotly>=5.17
tqdm>=4.66
scipy>=1.11
